Extract reading functions into RawGtfsReader
